Clean: Relocate STEX TraverseBorders
