+++ src/imageproc/TraverseBorders.h
